refactor: 환율 관련 클래스 패키지 위치 수정
