+++ src/main/java/shopping/order/service/currency/ExchangeRateProvider.java
@@ -1,4 +1,4 @@
-package shopping.currency;
+package shopping.order.service.currency;
 
 import shopping.order.domain.vo.ExchangeRate;
