+++ src/test/java/shopping/currency/MockExchangeRateProvider.java
@@ -3,6 +3,7 @@ package shopping.currency;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 import shopping.order.domain.vo.ExchangeRate;
+import shopping.order.service.currency.ExchangeRateProvider;
 
 @Profile("test")
 @Component
