+++ src/test/java/shopping/order/service/OrderServiceTest.java
@@ -15,7 +15,7 @@ import shopping.auth.domain.LoggedInMember;
 import shopping.cart.domain.CartItem;
 import shopping.cart.domain.vo.Quantity;
 import shopping.cart.repository.CartItemRepository;
-import shopping.currency.ExchangeRateProvider;
+import shopping.order.service.currency.ExchangeRateProvider;
 import shopping.exception.WooWaException;
 import shopping.member.domain.Member;
 import shopping.member.repository.MemberRepository;
